SharePoint connector: BigInteger size_bytes + per-row commit isolation
First sync surfaced two bugs:
1. SharepointDocument.size_bytes was Integer (32-bit, max ~2.1 GB) in
   the SQLAlchemy model but BigInteger in the migration. Some files
   in AMW SharePoint exceed 2 GB. Switched the model to BigInteger.
2. The connector batched commits per drive — when a single row
   failed (e.g. integer overflow), the whole transaction was poisoned
   and every subsequent row in the batch was skipped via a
   PendingRollbackError. Switched to per-row commit-or-rollback so
   one bad row doesn't tank the rest of the site.

+++ spider/apps/spider-kpi/backend/app/models/entities.py
@@ -4,7 +4,7 @@ from datetime import date, datetime
 from typing import Optional
 import uuid
 
-from sqlalchemy import ARRAY, Boolean, Date, DateTime, Float, ForeignKey, Index, Integer, Numeric, String, Text, UniqueConstraint, func
+from sqlalchemy import ARRAY, BigInteger, Boolean, Date, DateTime, Float, ForeignKey, Index, Integer, Numeric, String, Text, UniqueConstraint, func
 from sqlalchemy.dialects.postgresql import JSONB
 from sqlalchemy.orm import Mapped, mapped_column
 
@@ -1687,7 +1687,7 @@ class SharepointDocument(TimestampMixin, Base):
     path: Mapped[str] = mapped_column(String(2048), nullable=False, index=True)
     is_folder: Mapped[bool] = mapped_column(Boolean, default=False, nullable=False)
     top_level_folder: Mapped[Optional[str]] = mapped_column(String(255), index=True)
-    size_bytes: Mapped[Optional[int]] = mapped_column(Integer)
+    size_bytes: Mapped[Optional[int]] = mapped_column(BigInteger)
     mime_type: Mapped[Optional[str]] = mapped_column(String(128))
     web_url: Mapped[Optional[str]] = mapped_column(String(2048))
     created_by_email: Mapped[Optional[str]] = mapped_column(String(255))
